bin/server/pms/requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-extensions==2.1.4
djangorestframework==3.9.0
pkg-resources==0.0.0
pytz==2018.7
six==1.12.0
